pyproject.toml
[tool.poetry]
name = "programmable-vi-pldi-2024"
version = "0.0.1"
description = "The evaluation artifact for Probabilistic Programming with Programmable Variational Inference."
authors = ["McCoy Reynolds Becker <mccoyb@mit.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
adevjax = {path = "adevjax"}
genjax = {path = "genjax", develop = true}
python = ">=3.9,<3.13"
jax = "0.4.24"
jaxlib="0.4.24"
numpy = "1.26.4"
tensorflow-probability = "0.21.0"
equinox = "0.11.3"
numpyro = "^0.13.2"
torch = "2.1.0"
torchvision = "0.16.0"
pytest = "^8.0.2"
pytest-benchmark = "^4.0.0"
pyro-ppl = "1.8.6"
matplotlib = "^3.6.2"
seaborn = "^0.12.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv"
]
extend-include = ["*.ipynb"]
line-length = 88
indent-width = 4

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
extend-select = ["I"]
select = ["E4", "E7", "E9", "F"]
# F811 silences errors from GenJAX's type-based dispatch. Remove this if we can
# find some way to teach ruff about these overloads.
#
# F403 disables errors from `*` imports, which we currently use heavily.
ignore = ["F811", "F403"]
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
